--- textbox/model/unilm_v1/__init__.py
@@ -1,6 +1,0 @@
-__version__ = "0.4.0"
-from .tokenization_unilm import BertTokenizerForUnilm, BasicTokenizer, WordpieceTokenizer
-from .modeling import (BertConfig, BertModel, BertForPreTraining, BertForMaskedLM, BertForNextSentencePrediction, BertForSequenceClassification,
-                       BertForMultipleChoice, BertForTokenClassification, BertForQuestionAnswering, BertForPreTrainingLossMask, BertPreTrainingPairRel, BertPreTrainingPairTransform)
-from .optimization import BertAdam, BertAdamFineTune
-from .file_utils import PYTORCH_PRETRAINED_BERT_CACHE
