+++ transformers/src/transformers/utils/hub.py
@@ -39,7 +39,7 @@ from filelock import FileLock
 from huggingface_hub import HfFolder, Repository, create_repo, list_repo_files, whoami
 from requests.exceptions import HTTPError
 from requests.models import Response
-from transformers.utils.logging import tqdm
+from .logging import tqdm
 
 from . import __version__, logging
 from .import_utils import (
