+++ transformers/src/transformers/utils/import_utils.py
@@ -28,7 +28,7 @@ from typing import Any
 
 from packaging import version
 
-from transformers.utils.versions import importlib_metadata
+from .versions import importlib_metadata
 
 from . import logging
